requirements.txt
streamlit>=1.32.0
firebase-admin>=6.4.0
groq>=0.4.0
plotly>=5.18.0
pandas>=2.2.0
python-dotenv>=1.0.0
altair>=5.2.0
PyJWT>=2.8.0 